+++ src/piddiplatsch/commands/base.py
@@ -88,6 +88,8 @@ class KafkaCommand(Command, ABC):
         publish: bool = False,
         force: bool = False,
         idle_timeout: float | None = None,
+        limit: int | None = None,
+        monitor_db: bool = False,
         handle_profile: str | None = None,
     ) -> None:
         progress = self.progress(title=title, stream=True)
@@ -103,6 +105,8 @@ class KafkaCommand(Command, ABC):
                 publish=publish,
                 force=force,
                 idle_timeout=idle_timeout,
+                limit=limit,
+                monitor_db=monitor_db,
                 handle_profile=handle_profile,
             )